refactor: move getConsolidated to core +
- move getConsolidated to core
- reorg OS code on core

Ref eng-268
