+++ packages/core/src/command/consolidated/search/search-consolidated.ts
@@ -1,6 +1,6 @@
-import { Patient } from "@metriport/core/domain/patient";
-import { searchLexical } from "@metriport/core/external/opensearch/lexical/search";
 import { SearchSetBundle } from "@metriport/shared/medical";
+import { Patient } from "../../../domain/patient";
+import { searchLexical } from "./search-lexical";
 import { getConsolidatedPatientData } from "../consolidated-get";
 
 export async function searchConsolidated({
