+++ packages/core/src/external/opensearch/semantic/hybrid-create.ts
@@ -1,0 +1,30 @@
+// TODO eng-41 Convert this to actual code to create the index if not exists and decide where to call it from
+
+/*
+PUT /medical-resources
+{
+    "settings": {
+        "index.knn": true,
+        "default_pipeline": "embedding-ingest-pipeline",
+        "index.search.default_pipeline": "hybrid-search-pipeline"
+    },
+    "mappings": {
+        "properties": {
+            "content_embedding": {
+                "type": "knn_vector",
+                "dimension": 768,
+                "method": {
+                    "name": "hnsw",
+                    "space_type": "innerproduct",
+                    "engine": "nmslib"
+                }
+            },
+            "cxId": { "type": "keyword" },
+            "patientId": { "type": "keyword" },
+            "resourceType": { "type": "keyword" },
+            "resourceId": { "type": "keyword" },
+            "content": { "type": "text" }
+        }
+    }
+}
+*/
